removed build errors
Updated BroomPlayerController. 
during build Unity throws an error: "GetBool is not allowed to be called from a MonoBehaviour constructor  (or instance field initializer), call it from Awake or Start instead."

this error is due to initializing a variable to a Class/Object (etc.) (which in this case was TimerLogic) in constructor itself. Shifted it to Start method! And bang, Done!